requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
Authlib==0.14.3
autopep8==1.5.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
Django==3.1.1
dnspython==1.16.0
email-validator==1.1.1
Fabric3==1.14.post1
fastapi==0.62.0
flasgger==0.9.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-GraphQL==2.0.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
graphene==2.1.8
graphene-django==2.14.0
graphene-mongo==0.2.13
graphene-pydantic==0.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==20.0.4
h11==0.11.0
h2==4.0.0
hpack==4.0.0
httptools==0.1.1
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
iso8601==0.1.13
isort==5.4.2
isoweek==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.20.0
mysqlclient==2.0.1
paramiko==2.7.1
priority==1.3.0
promise==2.3
protobuf==3.12.4
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pylint==2.6.0
pymongo==3.11.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.13.6
text-unidecode==1.3
toml==0.10.1
urllib3==1.26.2
uvicorn==0.13.1
uvloop==0.14.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.0.0
wtf==0.1
WTForms==2.3.3
